Use one thread for every email
